+++ src/components/Activity/PopularActivityItem.tsx
@@ -2,7 +2,7 @@ import Star from "@/assets/svgs/star.svg";
 import Loading from "@/src/components/Loading";
 import PATH_NAMES from "@/src/constants/pathname";
 import useResponsiveTextStyle from "@/src/hooks/Activity/useResponsiveTextStyle";
-import useLoadingStore from "@/src/stores/loadingStore";
+import useLoadingStore from "@/src/stores/useLoadingStore";
 import type { PopularActivityItemProps } from "@/src/types/activity";
 import clsx from "clsx";
 import Image from "next/image";
